Fix svn:eol-style native for source files. Make a refed copy of the wxIcon from the wxIconBundle.
git-svn-id: https://svn.code.sf.net/p/wxlua/svn/trunk@81 bc88403a-3a91-4506-9056-abb4039e9f02

+++ wxLua/apps/wxlua/src/lconsole.cpp
@@ -231,7 +231,7 @@ void wxLuaConsole::DisplayStack(const wxLuaState& wxlState)
 
             // skip over ourselves on the stack
             if (nIndex > 0)
-                m_debugListBox->Append(buffer, (void *) nIndex);
+                m_debugListBox->Append(buffer, (void *)(long) nIndex);
         }
         nIndex++;
     }
